worker/requirements-worker.txt
torch==2.2.0+cu121
torchvision==0.17.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
diffusers==0.24.0
transformers==4.36.2
accelerate==0.25.0
safetensors==0.4.0
huggingface-hub==0.20.3
xformers==0.0.23.post1
opencv-python==4.8.1.78
imageio==2.31.0
imageio-ffmpeg==0.4.9
Pillow==10.1.0
httpx==0.25.2
boto3==1.34.0
